+++ library/core/src/main/java/com/google/android/exoplayer2/util/Util.java
@@ -1092,12 +1092,12 @@ public final class Util {
   }
 
   /**
-   * Returns whether {@code encoding} is one of the PCM encodings.
+   * Returns whether {@code encoding} is one of the linear PCM encodings.
    *
    * @param encoding The encoding of the audio data.
    * @return Whether the encoding is one of the PCM encodings.
    */
-  public static boolean isEncodingPcm(@C.Encoding int encoding) {
+  public static boolean isEncodingLinearPcm(@C.Encoding int encoding) {
     return encoding == C.ENCODING_PCM_8BIT
         || encoding == C.ENCODING_PCM_16BIT
         || encoding == C.ENCODING_PCM_24BIT